some linux fixes
git-svn-id: file:///home/nick/svnrepo/repo/indigo/trunk@945 20aa11e4-6e6f-594c-a032-6291ace9ee65

+++ utils/platformutils.h
@@ -88,7 +88,7 @@ void openFileBrowserWindowAtLocation(const std::string& select_path);
 const std::string getLastErrorString();
 
 
-uint32 getProcessID();
+uint64 getProcessID();
 
 
 enum ProcessPriority
